pyproject.toml
[project]
name = "listenlabs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "pyside6",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "ruff>=0.12.12",
    "scipy>=1.16.1",
]

[dependency-groups]
dev = [
    "ty>=0.0.1a20",
]


